Remap bond indices for CXSMILES

+++ umol-graph/src/io/smiles/parser/tests/utils.rs
@@ -226,7 +226,7 @@ pub fn build_from_graph(spec: &str) -> Molecule {
 
     let mut b = MoleculeBuilder::with_capacity(atoms.len(), bonds.len());
     // map of insertion index to atom id is identity by construction; still collect ids
-    let mut ids: Vec<u32> = Vec::with_capacity(atoms.len());
+    let mut ids: Vec<usize> = Vec::with_capacity(atoms.len());
     // Keep a map from atom span_start -> span_end for bond defaulting
     let mut atom_span_map: HashMap<u32, u32> = HashMap::new();
 
@@ -300,7 +300,7 @@ pub fn build_extended_from_graph(spec: &str) -> ExtendedMolecule {
     let bonds: Vec<_> = bonds_s.split_whitespace().collect();
 
     let mut b = ExtendedMoleculeBuilder::with_capacity(atoms.len(), bonds.len());
-    let mut ids: Vec<u32> = Vec::with_capacity(atoms.len());
+    let mut ids: Vec<usize> = Vec::with_capacity(atoms.len());
     let mut atom_span_map: HashMap<u32, u32> = HashMap::new();
 
     for tok in atoms {
